chore: restructure bench tests
