+++ .travis.yml
@@ -21,22 +21,30 @@ matrix:
   - name: "Python 2.7 Production Setup"
     python: 2.7
     env: TEST=bench
-    script: python -m unittest -v bench.tests.test_setup_production
+    script:
+      - python -m unittest -v bench.tests.test_init
+      - python -m unittest -v bench.tests.test_setup_production
 
   - name: "Python 3.6 Production Setup"
     python: 3.6
     env: TEST=bench
-    script: python -m unittest -v bench.tests.test_setup_production
+    script:
+      - python -m unittest -v bench.tests.test_init
+      - python -m unittest -v bench.tests.test_setup_production
 
   - name: "Python 3.7 Production Setup"
     python: 3.7
     env: TEST=bench
-    script: python -m unittest -v bench.tests.test_setup_production
+    script:
+      - python -m unittest -v bench.tests.test_init
+      - python -m unittest -v bench.tests.test_setup_production
 
   - name: "Python 3.8 Production Setup"
     python: 3.8
     env: TEST=bench
-    script: python -m unittest -v bench.tests.test_setup_production
+    script:
+      - python -m unittest -v bench.tests.test_init
+      - python -m unittest -v bench.tests.test_setup_production
 
   - name: "Python 3.6 Easy Install"
     python: 3.6
